Configures automated NPM publishing
Sets up GitHub Actions workflow to automatically publish the package to NPM on releases and pushes to the main branch.

This allows for both stable and beta releases to be published automatically, improving the release process and ensuring that the package is always up-to-date on NPM.

Also updates the README with instructions for installing stable and beta versions.
Adds .npmignore to exclude unnecessary files from the published package.
Adds metadata to package.json.